+++ frontend/components/play-bar.tsx
